+++ motion_control.h
@@ -29,9 +29,9 @@ void mc_init();
 // Execute linear motion in absolute millimeter coordinates. Feed rate given in millimeters/second
 // unless invert_feed_rate is true. Then the feed_rate means that the motion should be completed in
 // (1 minute)/feed_rate time.
-void mc_line(double x, double y, double z, float feed_rate, int invert_feed_rate, int16_t line_number);
+void mc_line(double x, double y, double z, float feed_rate, int invert_feed_rate, int32_t line_number);
 
-void mc_reposition(double x, double y, double z, int16_t line_number);
+void mc_reposition(double x, double y, double z, int32_t line_number);
 
 // Execute an arc. theta == start angle, angular_travel == number of radians to go along the arc,
 // positive angular_travel means clockwise, negative means counterclockwise. Radius == the radius of the
@@ -40,15 +40,15 @@ void mc_reposition(double x, double y, double z, int16_t line_number);
 void mc_arc(double theta, double angular_travel, double radius, double linear_travel, 
             int axis_1, int axis_2, int axis_linear, 
             double target_x, double target_y, double target_z,
-            double feed_rate, int invert_feed_rate, int16_t line_number);
+            double feed_rate, int invert_feed_rate, int32_t line_number);
 
 void mc_continue_arc();
 
 // Dwell for a couple of time units
-void mc_dwell(uint32_t milliseconds, int16_t line_number);
+void mc_dwell(uint32_t milliseconds, int32_t line_number);
 
 // Send the tool home
-void mc_go_home(int16_t line_number);
+void mc_go_home(int32_t line_number);
 
 // Stop any current motions
 void mc_stop();
@@ -60,7 +60,7 @@ void mc_stop();
 //void mc_continue(int line_number);
 
 extern volatile char mc_running;
-extern int16_t acting_line_number;
+extern int32_t acting_line_number;
 
 uint8_t mc_in_arc();
